Refactoring code + adding some updates in block add friend

+++ transcadence/api/views.py
@@ -5,7 +5,7 @@ from rest_framework.decorators import api_view
 from django.http import HttpResponse
 from rest_framework.views import APIView
 from .serializer import ChatsSerializer, MessageSerializer,InvitationSerializer
-from .models import Message,Invitations, Invitations
+from .models import Message,Invitations
 from django.db.models import Q
 
 
@@ -13,10 +13,15 @@ from django.db.models import Q
 def inviteFriend(request):
     serializer = InvitationSerializer(data=request.data)
     if serializer.is_valid():
-        serializer.save()
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-    else:
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        user1 = serializer.validated_data.get("user1")
+        user2 = serializer.validated_data.get("user2")
+        try: 
+            o = Invitations.objects.get(Q(user1=user1,user2=user2) | Q(user1=user2,user2=user1))
+            return Response("cant invite the player", status=status.HTTP_400_BAD_REQUEST)
+        except:
+            serializer.save()
+            return Response("Invited player successfuly", status=status.HTTP_201_CREATED)
+    return Response("cant invite the player", status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['POST'])
 def acceptFriend(request):
@@ -37,12 +42,17 @@ def acceptFriend(request):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
 @api_view(['POST'])
-def declineFriend(request, user1=None, user2=None):
-    try:
-        query = Invitations.objects.get(user2=user1,user1=user2,status="pending")
-        query.delete()
-    except:
-        return Response("Detail: Invitation Not found", status=status.HTTP_400_BAD_REQUEST)
+def declineFriend(request):
+    serializer= InvitationSerializer(data=request.data)
+    if (serializer.is_valid()):
+        validated_data = serializer.validated_data
+        user1 = validated_data.get('user1')
+        user2 = validated_data.get('user2')
+        try:
+            query = Invitations.objects.get(user2=user1,user1=user2,status="pending")
+            query.delete()
+        except:
+            return Response("Detail: Invitation Not found", status=status.HTTP_400_BAD_REQUEST)
     return Response("Detail: Declined successfully",status=status.HTTP_200_OK)
 
 @api_view(['POST'])
